requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
mysqlclient==2.2.1
Flask-Migrate==3.1.0
Flask-WTF==1.2.1
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==1.4.49
marshmallow==3.20.1
WTForms==3.0.1
pytest==7.4.2
pytest-flask==1.2.0
email-validator==2.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
bcrypt==4.0.1
Flask-Login==0.6.3
